Extracted LineAgg to a separate mod, mostly as is (except for visibility)
